requirements.txt
# LLM (OpenAI-compatible para GLM-4.7)
langchain-openai>=0.2.5
langchain-core>=0.3.0
openai>=1.50.0

# Job Search (sin MCP server)
python-jobspy>=1.1.4

# Google APIs
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Web framework + scheduler
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
apscheduler>=3.10.4

# HTTP
httpx>=0.27.2
requests>=2.31.0
aiohttp>=3.11.6

# Data
pydantic>=2.9.2
pydantic-settings>=2.5.2
python-dotenv>=1.0.1

# Utils
loguru>=0.7.2
rich>=13.9.2
